Sécurité, sql, diagramme de classes

+++ controller/AdminController.php
@@ -3,42 +3,76 @@ class AdminController
 {
     public function creerProduit()
     {
+        if (isset($_SESSION['user_id'])) {
+            if ($_SESSION['role'] == 'admin') {
+                if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+                    $produit = new ProduitRepository;
+                    try {
+                        $produit->createProduit($_POST['nomProduit'], $_POST['stockProduit']);
+                    } catch (Exception $ex) {
+                        echo $ex;
+                    }
 
-        if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-            $produit = new ProduitRepository;
-            $produit->createProduit($_POST['nomProduit'], $_POST['stockProduit']);
+                    header("Location: index.php?page=afficherproduits");
+                }
 
-            header("Location: index.php?page=afficherproduits");
+                require('view/creerProduits.php');
+            } else {
+                header("Location: index.php");
+            }
+        } else {
+            header("Location: index.php");
         }
-
-        require('view/creerProduits.php');
     }
 
     public function modifierProduit()
     {
-        if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+        if (isset($_SESSION['user_id'])) {
+            if ($_SESSION['role'] == 'admin') {
+                if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 
-            $produit = new ProduitRepository;
+                    $produit = new ProduitRepository;
 
-            $produit->modifierProduit($_POST['id'], $_POST['nomProduit'], $_POST['stockProduit']);
+                    try {
+                        $produit->modifierProduit($_POST['id'], $_POST['nomProduit'], $_POST['stockProduit']);
+                    } catch (Exception $ex) {
+                        echo $ex;
+                    }
 
-            header("Location: index.php?page=afficherproduits");
+                    header("Location: index.php?page=afficherproduits");
+                }
+                $produitRepo = new ProduitRepository;
+                $produitModifie = $produitRepo->getProduitById($_GET['id']);
+
+                require('view/modifierProduit.php');
+            } else {
+                header("Location: index.php");
+            }
+        } else {
+            header("Location: index.php");
         }
-        $produitRepo = new ProduitRepository;
-        $produitModifie = $produitRepo->getProduitById($_GET['id']);
-
-        require('view/modifierProduit.php');
     }
 
     public function supprimerProduit()
     {
-        if (isset($_GET['id'])) {
+        if (isset($_SESSION['user_id'])) {
+            if ($_SESSION['role'] == 'admin') {
+                if (isset($_GET['id'])) {
 
-            $produit = new ProduitRepository;
+                    $produit = new ProduitRepository;
+                    try {
+                        $produit->supprimerProduit($_GET['id']);
+                    } catch (Exception $ex) {
+                        echo $ex;
+                    }
 
-            $produit->supprimerProduit($_GET['id']);
-
-            header("Location: index.php?page=afficherproduits");
+                    header("Location: index.php?page=afficherproduits");
+                }
+            } else {
+                header("Location: index.php");
+            }
+        } else {
+            header("Location: index.php");
         }
     }
